refactor: MoveStrategy 삭제
 Piece와 통합
 패키지 정리
