+++ src/main/java/chess/domain/piece/King.java
@@ -1,0 +1,15 @@
+package chess.domain.piece;
+
+import chess.domain.position.MovableAreaFactory;
+import chess.domain.position.Position;
+
+public class King extends Piece {
+    public King(Position position) {
+        super(position, "k");
+    }
+
+    @Override
+    protected boolean isNotMovableTo(Position start, Position destination) {
+        return !MovableAreaFactory.kingOf(start).contains(destination);
+    }
+}
