+++ src/main/java/chess/domain/piece/Knight.java
@@ -1,0 +1,15 @@
+package chess.domain.piece;
+
+import chess.domain.position.MovableAreaFactory;
+import chess.domain.position.Position;
+
+public class Knight extends Piece {
+    public Knight(Position position) {
+        super(position, "n");
+    }
+
+    @Override
+    protected boolean isNotMovableTo(Position start, Position destination) {
+        return !MovableAreaFactory.knightOf(start).contains(destination);
+    }
+}
