--- src/main/java/chess/domain/piece/MoveStrategy.java
@@ -1,7 +1,0 @@
-package chess.domain.piece;
-
-import chess.domain.position.Position;
-
-public interface MoveStrategy {
-    boolean isNotMovableTo(Position start, Position destination);
-}
