+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,19 +1,23 @@
-package chess.domain.piece.queen;
+package chess.domain.piece;
 
-import chess.domain.piece.MoveStrategy;
 import chess.domain.position.MovableAreaFactory;
 import chess.domain.position.Position;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class QueenMoveStrategy implements MoveStrategy {
+public class Queen extends Piece {
+    public Queen(Position position) {
+        super(position, "q");
+    }
+
     @Override
-    public boolean isNotMovableTo(Position start, Position destination) {
+    protected boolean isNotMovableTo(Position start, Position destination) {
         List<Position> movable = new ArrayList<>();
         movable.addAll(MovableAreaFactory.columnOf(start));
         movable.addAll(MovableAreaFactory.rowOf(start));
         movable.addAll(MovableAreaFactory.diagonalsOf(start));
+
         return !movable.contains(destination);
     }
 }
