+++ src/main/java/chess/domain/piece/Rook.java
@@ -1,18 +1,22 @@
-package chess.domain.piece.rook;
+package chess.domain.piece;
 
-import chess.domain.piece.MoveStrategy;
 import chess.domain.position.MovableAreaFactory;
 import chess.domain.position.Position;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class RookMoveStrategy implements MoveStrategy {
+public class Rook extends Piece {
+    public Rook(Position position) {
+        super(position, "r");
+    }
+
     @Override
-    public boolean isNotMovableTo(Position start, Position destination) {
+    protected boolean isNotMovableTo(Position start, Position destination) {
         List<Position> movableArea = new ArrayList<>();
         movableArea.addAll(MovableAreaFactory.columnOf(start));
         movableArea.addAll(MovableAreaFactory.rowOf(start));
+
         return !movableArea.contains(destination);
     }
 }
