--- src/main/java/chess/domain/piece/bishop/Bishop.java
@@ -1,10 +1,0 @@
-package chess.domain.piece.bishop;
-
-import chess.domain.piece.Piece;
-import chess.domain.position.Position;
-
-public class Bishop extends Piece {
-    public Bishop(Position position) {
-        super(position, "b", new BishopMoveStrategy());
-    }
-}
