--- src/main/java/chess/domain/piece/knight/Knight.java
@@ -1,10 +1,0 @@
-package chess.domain.piece.knight;
-
-import chess.domain.piece.Piece;
-import chess.domain.position.Position;
-
-public class Knight extends Piece {
-    public Knight(Position position) {
-        super(position, "n", new KnightMoveStrategy());
-    }
-}
