--- src/main/java/chess/domain/piece/queen/Queen.java
@@ -1,10 +1,0 @@
-package chess.domain.piece.queen;
-
-import chess.domain.piece.Piece;
-import chess.domain.position.Position;
-
-public class Queen extends Piece {
-    public Queen(Position position) {
-        super(position, "q", new QueenMoveStrategy());
-    }
-}
