--- src/main/java/chess/domain/piece/rook/Rook.java
@@ -1,10 +1,0 @@
-package chess.domain.piece.rook;
-
-import chess.domain.piece.Piece;
-import chess.domain.position.Position;
-
-public class Rook extends Piece {
-    public Rook(Position position) {
-        super(position, "r", new RookMoveStrategy());
-    }
-}
