+++ src/test/java/chess/domain/piece/BishopTest.java
@@ -1,29 +1,26 @@
-package chess.domain.piece.pawn;
+package chess.domain.piece;
 
-import chess.domain.piece.Piece;
 import org.junit.jupiter.api.Test;
 
 import static chess.domain.position.Fixtures.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
-class PawnTest {
+class BishopTest {
 
     @Test
     void moveTo_When_Success() {
-        Piece pawn = new Pawn(C2);
-        pawn.moveTo(C4);
-        pawn.moveTo(C5);
-        pawn.moveTo(B6);
+        Piece bishop = new Bishop(C3);
+        bishop.moveTo(H8);
 
-        assertThat(pawn.getPosition()).isEqualTo(B6);
+        assertThat(bishop.getPosition()).isEqualTo(H8);
     }
 
     @Test
     void moveTo_When_Fail() {
-        Piece pawn = new Pawn(C3);
+        Piece bishop = new Bishop(C3);
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> pawn.moveTo(C5))
+                .isThrownBy(() -> bishop.moveTo(D3))
                 .withMessage("기물의 이동 범위에 속하지 않습니다.");
     }
 }
