+++ src/test/java/chess/domain/piece/QueenTest.java
@@ -1,28 +1,27 @@
-package chess.domain.piece.knight;
+package chess.domain.piece;
 
-import chess.domain.piece.Piece;
 import org.junit.jupiter.api.Test;
 
 import static chess.domain.position.Fixtures.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
-public class KnightTest {
+class QueenTest {
 
     @Test
     void moveTo_When_Success() {
-        Piece knight = new Knight(C3);
-        knight.moveTo(D1);
-        knight.moveTo(B2);
+        Piece queen = new Queen(C3);
+        queen.moveTo(H8);
+        queen.moveTo(H3);
 
-        assertThat(knight.getPosition()).isEqualTo(B2);
+        assertThat(queen.getPosition()).isEqualTo(H3);
     }
 
     @Test
     void moveTo_When_Fail() {
-        Piece knight = new Knight(C3);
+        Piece queen = new Queen(C3);
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> knight.moveTo(D3))
+                .isThrownBy(() -> queen.moveTo(D1))
                 .withMessage("기물의 이동 범위에 속하지 않습니다.");
     }
 }
